ANDROID: Fix up KABI breakage in 4.19.209 in struct sock
Commit 35306eb23814 ("af_unix: fix races in sk_peer_pid and sk_peer_cred
accesses") was backported to 4.19.209 and it adds a new field to struct
sock.  We want to keep this commit, so rework it using the reserved
fields we have set up for this very issue.

However, putting a spinlock_t into a 64bit value breaks when building
with some debugging options enabled, so we have to disable them when
doing build testing for the allmodconfig builds.  In order to do that,
the following config options are forced off:
	DEBUG_SPINLOCK
	DEBUG_LOCK_ALLOC
	LOCK_STAT
	DEBUG_WW_MUTEX_SLOWPATH
	DEBUG_LOCK_ALLOC
	PROVE_LOCKING

We only really want to disable DEBUG_SPINLOCK and DEBUG_LOCK_ALLOC as
those affect the size of spinlock_t, but the other options need to be
disabled as well because they try to force on the locking debug config
options when selected.

Note, this does not affect any device's behavior, those options are only
being disabled for the test systems when doing allmodconfig builds.
They are not enabled in the normal GKI builds, so all is fine there.

Bug: 161946584
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I23df70a6bcbbc9d773e790cdf0377c1c67cdf682