defconfig: Disable touchscreen heatmap to reduce latency
Despite redbull having a 10 MHz SPI bus for the touchscreen, transfering
heatmap data at 120 Hz is still expensive and adds latency. It also
wastes power by keeping the bus active for longer than necessary.

The deep press feature that uses heatmaps doesn't work very well anyway,
so let's just disable it.

An interesting side effect of the heatmap feature is that it increases
jitter in touch readings. Drawing curves with a single finger results in
slightly jittery lines when it is enabled, and while there is still a
little bit of jitter with it disabled, the consistency of the lines is
better overall.

Test: touch works normally
Test: latency and jitter in Touchpaint app
Signed-off-by: Danny Lin <danny@kdrag0n.dev>